+++ src/main/java/com/batchproject/jobs/models/SysDataRepository.java
@@ -4,5 +4,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface SysDataRepository extends JpaRepository<SysData, Integer > {
+public interface SysDataRepository extends JpaRepository<SysData, Long > {
 }
