+++ src/main/java/com/batchproject/jobs/models/housing/HousingBuildingRepository.java
@@ -2,6 +2,6 @@ package com.batchproject.jobs.models.housing;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface HousingBuildingRepository extends JpaRepository<HousingBuilding, Integer> {
+public interface HousingBuildingRepository extends JpaRepository<HousingBuilding, Long> {
 
 }
