+++ src/main/java/com/batchproject/jobs/models/housing/SuiteDTO.java
@@ -7,12 +7,12 @@ import java.time.LocalDate;
 @Data
 public class SuiteDTO {
 
-    private Integer addressId;
+    private Long addressId;
     private LocalDate builtOn;
     private LocalDate lastRenovatedOn;
     private Integer noOfBedRooms;
     private Integer noOfBathRooms;
     private Boolean haveDedicatedLaundry;
     private Integer floorNo;
-    private Integer buildingId;
+    private Long buildingId;
 }
