+++ src/main/java/com/batchproject/jobs/models/maintenance/MaintenanceRepository.java
@@ -2,5 +2,5 @@ package com.batchproject.jobs.models.maintenance;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface MaintenanceRepository extends JpaRepository<Maintenance,Integer> {
+public interface MaintenanceRepository extends JpaRepository<Maintenance,Long> {
 }
