+++ src/main/java/com/batchproject/jobs/models/rent/RentRepository.java
@@ -2,7 +2,7 @@ package com.batchproject.jobs.models.rent;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface RentRepository extends JpaRepository<Rent, Integer> {
+public interface RentRepository extends JpaRepository<Rent, Long> {
 }
