+++ src/main/java/com/batchproject/jobs/models/tenant/TenantProfileRepository.java
@@ -2,5 +2,5 @@ package com.batchproject.jobs.models.tenant;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface TenantProfileRepository extends JpaRepository<TenantProfile,Integer> {
+public interface TenantProfileRepository extends JpaRepository<TenantProfile,Long> {
 }
